+++ test/CMakeLists.txt
@@ -19,6 +19,6 @@ set(CMAKE_EXE_LINKER_FLAGS "-lgcov")
 
 ## Define executables for each test suite
 set(SIMPLE "runSimpleTests")
-add_executable(${SIMPLE} "test/simple_tests.cpp" "../src/vector1.cpp")
+add_executable(${SIMPLE} "simple_tests.cpp" "../src/vector.cpp")
